+++ cartridge/confapplier.lua
@@ -494,7 +494,10 @@ local function boot_instance(clusterwide_config)
         local leader = topology_cfg.servers[leader_uuid]
 
         -- Set up 'star' replication for the bootstrap
-        if instance_uuid == leader_uuid then
+        local bootstrap_from = require('cartridge.argparse').get_opts({bootstrap_from = 'string'}).bootstrap_from
+        if bootstrap_from ~= nil then
+            box_opts.replication = {bootstrap_from}
+        elseif instance_uuid == leader_uuid then
             box_opts.replication = nil
             box_opts.read_only = false
             -- leader should be bootstrapped with quorum = 0, otherwise
@@ -563,6 +566,13 @@ local function boot_instance(clusterwide_config)
 
     do
         local read_only = box.cfg.read_only
+        local user = box.space[box.schema.USER_ID].index.name:get(username)
+
+        local remote_control_suspended = false
+        if vars.upgrade_schema or (user == nil or user.auth['chap-sha1'] ~= box.schema.user.password(password)) then
+            remote_control.suspend()
+            remote_control_suspended = true
+        end
 
         if vars.upgrade_schema then
             log.info('Upgrading schema ...')
@@ -575,7 +585,7 @@ local function boot_instance(clusterwide_config)
         -- Function `passwd` is safe to be called on multiple replicas,
         -- it never cause replication conflict
         -- But don't commit anything if it's already ok.
-        local user = box.space[box.schema.USER_ID].index.name:get(username)
+
         -- https://github.com/tarantool/tarantool/blob/2.7.3/src/box/lua/schema.lua#L2719-L2724
         if user == nil or user.auth['chap-sha1'] ~= box.schema.user.password(password) then
             log.info('Setting password for user %q ...', username)
@@ -587,6 +597,9 @@ local function boot_instance(clusterwide_config)
         end
 
         box.cfg({read_only = read_only})
+        if remote_control_suspended then
+            remote_control.resume()
+        end
     end
 
     -- Box is ready, start listening full-featured iproto protocol